requirements.txt
# Dev environment
pip
autopep8

# App
flask

# Image processing and AI
google-generativeai
Pillow
